refactor: rename UploadNotificationController to generic EventNotificationController
Replaces the upload-specific notification controller and its nested
module path with a generic EventNotificationController in
controller/notification_controller.py, and moves the upload controller
to controller/upload_controller.py. Updates all imports and references
across __main__, plugin, and tests accordingly.
