+++ OTCamera/plugin/upload_notifier/rabbitmq_s3_upload_to_otcloud.py
@@ -2,8 +2,8 @@ import dataclasses
 import json
 
 from OTCamera.config import OTCloudSettings
-from OTCamera.controller.upload.notification.controller import PayloadFactory
-from OTCamera.domain.events import FileUploaded, S3FileUploaded
+from OTCamera.controller.notification_controller import PayloadFactory
+from OTCamera.domain.events import Event, S3FileUploaded
 from OTCamera.plugin.upload_notifier.dtos import CameraIdDto, S3FileUploadedEvent
 
 
@@ -12,8 +12,8 @@ class RabbitMQS3UploadToOTCloudPayloadFactory(PayloadFactory):
     def __init__(self, ot_cloud_settings: OTCloudSettings):
         self.ot_cloud_settings = ot_cloud_settings
 
-    def create(self, event: FileUploaded) -> str:
-        # The EventBus subscription in UploadNotificationController already
+    def create(self, event: Event) -> str:
+        # The EventBus subscription in EventNotificationController already
         # filters to S3FileUploaded events, so this assert should never fire.
         # It is kept as a safety net against accidental misconfiguration at
         # the wiring site (e.g. wrong event_type passed to the controller).
@@ -25,6 +25,7 @@ class RabbitMQS3UploadToOTCloudPayloadFactory(PayloadFactory):
                 project_id=self.ot_cloud_settings.project_id,
                 site_id=self.ot_cloud_settings.site_id,
             ),
+            # TODO: verify that timestamp includes UTC/TZ info
             timestamp=event.timestamp.isoformat(),
             original_filename=event.filename,
             new_filename=event.filename,
